+++ lib/linear_percent_indicator.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+@Deprecated('This property is no longer used, please use barRadius instead.')
 enum LinearStrokeCap { butt, round, roundAll }
 
 // ignore: must_be_immutable
@@ -41,7 +42,11 @@ class LinearPercentIndicator extends StatefulWidget {
   final Widget? center;
 
   ///The kind of finish to place on the end of lines drawn, values supported: butt, round, roundAll
+  @Deprecated('This property is no longer used, please use barRadius instead.')
   final LinearStrokeCap? linearStrokeCap;
+
+  /// The border radius of the progress bar (Will replace linearStrokeCap)
+  final Radius? barRadius;
 
   ///alignment of the Row (leading-widget-center-trailing)
   final MainAxisAlignment alignment;
@@ -102,6 +107,7 @@ class LinearPercentIndicator extends StatefulWidget {
     this.center,
     this.addAutomaticKeepAlive = true,
     this.linearStrokeCap,
+    this.barRadius,
     this.padding = const EdgeInsets.symmetric(horizontal: 10.0),
     this.alignment = MainAxisAlignment.start,
     this.maskFilter,
@@ -247,16 +253,16 @@ class _LinearPercentIndicatorState extends State<LinearPercentIndicator>
         children: [
           CustomPaint(
             key: _containerKey,
-            painter: LinearPainter(
+            painter: _LinearPainter(
               isRTL: widget.isRTL,
               progress: _percent,
               progressColor: widget.progressColor,
               linearGradient: widget.linearGradient,
               backgroundColor: widget.backgroundColor,
+              barRadius: widget.barRadius ??
+                  Radius.zero, // If radius is not defined, set it to zero
               linearGradientBackgroundColor:
                   widget.linearGradientBackgroundColor,
-              linearStrokeCap: widget.linearStrokeCap,
-              lineWidth: widget.lineHeight,
               maskFilter: widget.maskFilter,
               clipLinearGradient: widget.clipLinearGradient,
             ),
@@ -311,61 +317,49 @@ class _LinearPercentIndicatorState extends State<LinearPercentIndicator>
   bool get wantKeepAlive => widget.addAutomaticKeepAlive;
 }
 
-class LinearPainter extends CustomPainter {
+class _LinearPainter extends CustomPainter {
   final Paint _paintBackground = new Paint();
   final Paint _paintLine = new Paint();
-  final double lineWidth;
   final double progress;
   final bool isRTL;
   final Color progressColor;
   final Color backgroundColor;
-  final LinearStrokeCap? linearStrokeCap;
+  final Radius barRadius;
   final LinearGradient? linearGradient;
   final LinearGradient? linearGradientBackgroundColor;
   final MaskFilter? maskFilter;
   final bool clipLinearGradient;
 
-  LinearPainter({
-    required this.lineWidth,
+  _LinearPainter({
     required this.progress,
     required this.isRTL,
     required this.progressColor,
     required this.backgroundColor,
-    this.linearStrokeCap = LinearStrokeCap.butt,
+    required this.barRadius,
     this.linearGradient,
     this.maskFilter,
     required this.clipLinearGradient,
     this.linearGradientBackgroundColor,
   }) {
     _paintBackground.color = backgroundColor;
-    _paintBackground.style = PaintingStyle.stroke;
-    _paintBackground.strokeWidth = lineWidth;
 
     _paintLine.color = progress.toString() == "0.0"
         ? progressColor.withOpacity(0.0)
         : progressColor;
-    _paintLine.style = PaintingStyle.stroke;
-    _paintLine.strokeWidth = lineWidth;
-
-    if (linearStrokeCap == LinearStrokeCap.round) {
-      _paintLine.strokeCap = StrokeCap.round;
-    } else if (linearStrokeCap == LinearStrokeCap.butt) {
-      _paintLine.strokeCap = StrokeCap.butt;
-    } else {
-      _paintLine.strokeCap = StrokeCap.round;
-      _paintBackground.strokeCap = StrokeCap.round;
-    }
   }
 
   @override
   void paint(Canvas canvas, Size size) {
-    final start = Offset(0.0, size.height / 2);
-    final end = Offset(size.width, size.height / 2);
-    canvas.drawLine(start, end, _paintBackground);
+    // Draw background first
+    Path backgroundPath = Path();
+    backgroundPath.addRRect(RRect.fromRectAndRadius(
+        Rect.fromLTWH(0, 0, size.width, size.height), barRadius));
+    canvas.drawPath(backgroundPath, _paintBackground);
 
     if (maskFilter != null) {
       _paintLine.maskFilter = maskFilter;
     }
+
     if (linearGradientBackgroundColor != null) {
       Offset shaderEndPoint =
           clipLinearGradient ? Offset.zero : Offset(size.width, size.height);
@@ -373,19 +367,25 @@ class LinearPainter extends CustomPainter {
           ?.createShader(Rect.fromPoints(Offset.zero, shaderEndPoint));
     }
 
+    // Then draw progress line
+    final xProgress = size.width * progress;
+    Path linePath = Path();
     if (isRTL) {
-      final xProgress = size.width - size.width * progress;
       if (linearGradient != null) {
         _paintLine.shader = _createGradientShaderRightToLeft(size, xProgress);
       }
-      canvas.drawLine(end, Offset(xProgress, size.height / 2), _paintLine);
+      linePath.addRRect(RRect.fromRectAndRadius(
+          Rect.fromLTWH(
+              size.width - size.width * progress, 0, xProgress, size.height),
+          barRadius));
     } else {
-      final xProgress = size.width * progress;
       if (linearGradient != null) {
         _paintLine.shader = _createGradientShaderLeftToRight(size, xProgress);
       }
-      canvas.drawLine(start, Offset(xProgress, size.height / 2), _paintLine);
+      linePath.addRRect(RRect.fromRectAndRadius(
+          Rect.fromLTWH(0, 0, xProgress, size.height), barRadius));
     }
+    canvas.drawPath(linePath, _paintLine);
   }
 
   Shader _createGradientShaderRightToLeft(Size size, double xProgress) {
@@ -412,7 +412,5 @@ class LinearPainter extends CustomPainter {
   }
 
   @override
-  bool shouldRepaint(CustomPainter oldDelegate) {
-    return true;
-  }
+  bool shouldRepaint(CustomPainter oldDelegate) => true;
 }